editor/libeditor/tests/test_can_undo_after_setting_value.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin"
                 type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1386222
-->
<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Mozilla Bug 1386222" onload="runTest();">
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <input xmlns="http://www.w3.org/1999/xhtml" id="t"/>
  <script class="testbody" type="application/javascript">
  <![CDATA[

function runTest() {
  var t = document.getElementById("t");
  is(
    t.editor.canUndo,
    false,
    "Editor shouldn't have undo transaction at start"
  );
  t.value = "foo";
  is(t.value, "foo", "value setter worked");
  is(
    t.editor.canUndo,
    true,
    "Editor should have undo transaction after setting value"
  );
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
]]>
</script>
</window>
